added VS_ATTEN_CHAR and dSPACE SCS Diag message
