+++ Wireshark/profiles/dsV2Gshark/colorfilters
@@ -1,4 +1,4 @@
-@Homeplug@homeplug-av@[26214,26214,26214][65535,65535,65535]
+@Homeplug@homeplug-av or homeplug-scs@[26214,26214,26214][65535,65535,65535]
 @V2G Warning@v2gtp and _ws.expert@[52685,65535,51657][54484,0,0]
 @V2G Default@v2gtp@[52685,65535,51657][0,0,0]
 @V2G TLS Secret@v2gtlssecret@[0,29555,2056][65535,65535,65535]
